Use joblib to parallelize downloads

+++ download_images.py
@@ -6,6 +6,7 @@ import pandas as pd
 from argparse import ArgumentParser
 from urllib import request
 from tqdm import trange
+from joblib import Parallel, delayed
 
 parser = ArgumentParser()
 parser.add_argument(
@@ -29,17 +30,21 @@ if not os.path.isdir(args.output_dirpath):
     os.makedirs(args.output_dirpath)
 
 print('[Download images into "{}"]'.format(args.output_dirpath))
-for i in trange(len(csv)):
+
+def download_image(i):
     filename = csv.iloc[i]['Filename']
     category = csv.iloc[i]['Category']
     inner_output_dirpath = os.path.join(args.output_dirpath, category)
     if not os.path.isdir(inner_output_dirpath):
         os.mkdir(inner_output_dirpath)
     output_filepath = os.path.join(inner_output_dirpath, filename)
-    
+
     url = csv.iloc[i]['Image URL']
-    downloaded_img = request.urlopen(url)
-    f = open(output_filepath, mode='wb')
-    f.write(downloaded_img.read())
-    downloaded_img.close()
-    f.close()
+    if not os.path.isfile(output_filepath):
+        downloaded_img = request.urlopen(url)
+        f = open(output_filepath, mode='wb')
+        f.write(downloaded_img.read())
+        downloaded_img.close()
+        f.close()
+
+Parallel(n_jobs=-1)(delayed(download_image)(i) for i in trange(len(csv)))
